+++ HISTORY.md
@@ -5,6 +5,7 @@
 ## Unreleased
 - Improved Python API authentication when using URL-based connectivity,
   by respecting the `credential` keyword argument
+- Added basic example program about how to import a dashboard
 
 ## 0.4.0 (2024-10-16)
 - Fixed folder argument issue